Fix build issues under Clang
Most of these changes are trivial to satisfy Clang's warnings. Clang did
find one small bug in the move constructor for
Core::Debug::DebugMessage, where processName and threadName were not
moved over as they should have been.

Adds Clang v3.4-3.7 as supported compilers to the README and adds
instructions to CLANG.md.

Towards #9: port to clang

+++ Core/ProtoBuf.cc
@@ -166,9 +166,9 @@ serialize(const google::protobuf::Message& from,
               from.InitializationErrorString().c_str(),
               dumpString(from).c_str());
     }
-    uint32_t length = from.ByteSize();
+    uint32_t length = uint32_t(from.ByteSize());
     char* data = new char[skipBytes + length];
-    from.SerializeToArray(data + skipBytes, length);
+    from.SerializeToArray(data + skipBytes, int(length));
     to.setData(data, skipBytes + length, Core::Buffer::deleteArrayFn<char>);
 }
